+++ examples/generators/main_ids.py
@@ -5,6 +5,8 @@ import slangpy as spy
 import pathlib
 import numpy as np
 
+print("SlangPy id generator examples (https://slangpy.shader-slang.org/en/latest/generator_ids.html)")
+
 # Create an SGL device with the local folder for slangpy includes
 device = spy.create_device(include_paths=[
     pathlib.Path(__file__).parent.absolute(),