+++ examples/nested/main.py
@@ -1,8 +1,11 @@
-# SPDX-License-Identifier: Apache-2.0
+# SPDX-License-Identifier: Apache-2.0 WITH LLVM-exception
 
 import sgl
 import slangpy as spy
 import pathlib
+
+print("SlangPy nested types example (https://slangpy.shader-slang.org/en/latest/nested.html)")
+print("This example requires tev (https://github.com/Tom94/tev) to display results.")
 
 # Create an SGL device with the local folder for slangpy includes
 device = spy.create_device(include_paths=[