+++ slangpy/__init__.py
@@ -6,7 +6,7 @@ import runpy
 import pathlib
 
 # Version number
-__version__ = "0.21.0"
+__version__ = "0.21.1"
 
 # Useful slangpy types
 from . import types